Further updates for the assets folder cleanup.

+++ .github/_archive/actions/migrate_to_prod/migrate_prod.sh
@@ -53,10 +53,21 @@ if [ -n "$WORDPRESS_CONTAINER_NAME" ]; then
     # Copy over backup file
     oc cp --no-preserve ./wp-backup.wpress -n $NAMESPACE -c $WORDPRESS_CONTAINER_NAME $WORDPRESS_POD_NAME:/var/www/html/wp-content/ai1wm-backups/wp-backup.wpress
 
-    #erase any cleanbc plugin assets before restore
+    #erase any cleanbc plugin assets before restore    
     echo "Cleaning the bcgov-plugin-cleanbc/dist/assets folder"
-    oc exec  -n $NAMESPACE -c $WORDPRESS_CONTAINER_NAME $WORDPRESS_POD_NAME -- bash -c "rm /var/www/html/wp-content/plugins/bcgov-plugin-cleanbc/dist/assets/*"
+    set +e
+	CLEANBCPLUGIN_VERSION=$(oc exec -n $NAMESPACE -c $WORDPRESS_CONTAINER_NAME $WORDPRESS_POD_NAME -- php /tmp/wp-cli.phar plugin get bcgov-plugin-cleanbc --field=version 2>&1)
+	CLEANBCPLUGIN_VERSION_EXIT_CODE=$?
+	set -e
+	if [ CLEANBCPLUGIN_VERSION_EXIT_CODE -eq 0 ]; then		
+        oc exec  -n $NAMESPACE -c $WORDPRESS_CONTAINER_NAME $WORDPRESS_POD_NAME -- bash -c "rm /var/www/html/wp-content/plugins/bcgov-plugin-cleanbc/dist/assets/*"
+    
+	else
+		echo "bcgov-plugin-cleanbc Not installed"
 
+		echo "::warning::bcgov-plugin-cleanbc Not installed"
+	fi
+    
     #perform the restore
     oc exec -n $NAMESPACE -c $WORDPRESS_CONTAINER_NAME $WORDPRESS_POD_NAME -- bash -c "echo 'y' | php /tmp/wp-cli.phar ai1wm restore wp-backup.wpress"
